Refactor rendering methods
